feat: Add utility function to generate unique IDs
This commit adds a new utility function called `generateId` to the `get-id.ts` file in the `website/src/utilts` directory. The function uses the `uuidv7` library to generate a unique ID and caches the result using the `cache` function from the `react` package.

The purpose of this change is to provide a convenient way to generate unique IDs for various purposes in the website.